backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
sendgrid==6.11.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
pytesseract==0.3.10
apscheduler==3.10.4
redis==5.0.1
celery==5.3.6
feedparser==6.0.11
